backend/requirements.txt
fastapi
uvicorn
asyncpg
pydantic
pyyaml
pytest
httpx
pyserial-asyncio